requirements.txt
openai
langchain
langchain_core
langchain_community
streamlit
langchain_openai
python-dotenv
PyPDF2

#API Requirements
fastAPI
langserve
uvicorn
sse_starlette

# RAG Requirements
bs4
pypdf
chromadb
faiss-cpu


-e .

# Compatible Versions

# pip install fastapi==0.111.0 langserve==0.3.1 langchain==0.3.25 langchain-core==0.3.65 pydantic==2.7.4